Move durable opfs-btree WAL into core crate
