+++ crates/jazz-tools/src/storage/opfs_btree.rs
@@ -6,25 +6,27 @@
 //! Key encoding scheme (all keys are UTF-8 strings with hex-encoded binary parts):
 //!
 //! ```text
-//! "obj:{uuid}:meta"                                       → JSON metadata
-//! "obj:{uuid}:br:{branch}:tips"                           → JSON HashSet<CommitId>
-//! "obj:{uuid}:br:{branch}:c:{commit_uuid}"                → JSON Commit
-//! "ack:{commit_hex}"                                      → JSON HashSet<DurabilityTier>
-//! "catman:{app_uuid}:op:{object_uuid}"                    → JSON CatalogueManifestOp
-//! "idx:{table}:{col}:{branch}:{hex_encoded_value}:{uuid}" → empty (existence is the signal)
+//! "obj:{uuid}:meta"                                       -> JSON metadata
+//! "obj:{uuid}:br:{branch}:tips"                           -> JSON HashSet<CommitId>
+//! "obj:{uuid}:br:{branch}:c:{commit_uuid}"                -> JSON Commit
+//! "ack:{commit_hex}"                                      -> JSON HashSet<DurabilityTier>
+//! "catman:{app_uuid}:op:{object_uuid}"                    -> JSON CatalogueManifestOp
+//! "idx:{table}:{col}:{branch}:{hex_encoded_value}:{uuid}" -> empty (existence is the signal)
 //! ```
 
 use std::cell::RefCell;
 use std::collections::{HashMap, HashSet};
 use std::ops::Bound;
 #[cfg(not(target_arch = "wasm32"))]
-use std::path::{Path, PathBuf};
+use std::path::Path;
 
 #[cfg(target_arch = "wasm32")]
 use opfs_btree::OpfsFile;
 #[cfg(not(target_arch = "wasm32"))]
 use opfs_btree::StdFile;
-use opfs_btree::{BTreeError, BTreeOptions, MemoryFile, OpfsBTree, SyncFile};
+use opfs_btree::{
+    BTreeError, BTreeOptions, DurableBTreeFiles, DurableOpfsBTree, MemoryFile, SyncFile,
+};
 
 use crate::commit::{Commit, CommitId};
 use crate::object::{BranchName, ObjectId};
@@ -44,14 +46,6 @@ use super::{
 };
 
 const MIN_CACHE_SIZE_BYTES: usize = 4 * 1024 * 1024;
-const FULL_SCAN_END_KEY: [u8; 1] = [u8::MAX];
-const MANIFEST_MAGIC: [u8; 8] = *b"JAZZWAL1";
-const MANIFEST_VERSION: u32 = 1;
-const MANIFEST_SLOT_BYTES: usize = 128;
-const MANIFEST_CHECKSUM_BYTES: usize = 8;
-const WAL_BATCH_MAGIC: [u8; 8] = *b"JAZZWLB1";
-const WAL_BATCH_HEADER_BYTES: usize = 8 + 4 + MANIFEST_CHECKSUM_BYTES;
-type EncodedKv = (Vec<u8>, Vec<u8>);
 
 #[derive(Clone, Debug)]
 enum AnyFile {
@@ -114,280 +108,22 @@ impl SyncFile for AnyFile {
     }
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ReplicaSlot {
-    A,
-    B,
-}
-
-impl ReplicaSlot {
-    fn inactive(self) -> Self {
-        match self {
-            Self::A => Self::B,
-            Self::B => Self::A,
-        }
-    }
-
-    fn encode(self) -> u8 {
-        match self {
-            Self::A => 0,
-            Self::B => 1,
-        }
-    }
-
-    fn decode(raw: u8) -> Result<Self, StorageError> {
-        match raw {
-            0 => Ok(Self::A),
-            1 => Ok(Self::B),
-            other => Err(StorageError::IoError(format!(
-                "opfs-btree durable manifest: invalid slot {}",
-                other
-            ))),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ManifestSlot {
-    A,
-    B,
-}
-
-impl ManifestSlot {
-    fn inactive(self) -> Self {
-        match self {
-            Self::A => Self::B,
-            Self::B => Self::A,
-        }
-    }
-
-    fn byte_offset(self) -> u64 {
-        match self {
-            Self::A => 0,
-            Self::B => MANIFEST_SLOT_BYTES as u64,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct DurableManifest {
-    generation: u64,
-    active_snapshot: ReplicaSlot,
-    active_wal: ReplicaSlot,
-    applied_wal_seq: u64,
-}
-
-impl DurableManifest {
-    fn bootstrap() -> Self {
-        Self {
-            generation: 1,
-            active_snapshot: ReplicaSlot::A,
-            active_wal: ReplicaSlot::A,
-            applied_wal_seq: 0,
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum WalEntry {
-    Put {
-        seq: u64,
-        key: Vec<u8>,
-        value: Vec<u8>,
-    },
-    Delete {
-        seq: u64,
-        key: Vec<u8>,
-    },
-}
-
-impl WalEntry {
-    fn seq(&self) -> u64 {
-        match self {
-            Self::Put { seq, .. } | Self::Delete { seq, .. } => *seq,
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-struct DurableFiles {
-    snapshot_a: AnyFile,
-    snapshot_b: AnyFile,
-    wal_a: AnyFile,
-    wal_b: AnyFile,
-    manifest: AnyFile,
-}
-
-impl DurableFiles {
-    fn memory() -> Self {
-        Self {
-            snapshot_a: AnyFile::Memory(MemoryFile::new()),
-            snapshot_b: AnyFile::Memory(MemoryFile::new()),
-            wal_a: AnyFile::Memory(MemoryFile::new()),
-            wal_b: AnyFile::Memory(MemoryFile::new()),
-            manifest: AnyFile::Memory(MemoryFile::new()),
-        }
-    }
-
-    #[cfg(not(target_arch = "wasm32"))]
-    fn open(path: impl AsRef<Path>) -> Result<Self, StorageError> {
-        let path = path.as_ref();
-        Ok(Self {
-            snapshot_a: AnyFile::Std(StdFile::open(path).map_err(map_storage_err)?),
-            snapshot_b: AnyFile::Std(
-                StdFile::open(sidecar_path(path, ".snapshot-b")).map_err(map_storage_err)?,
-            ),
-            wal_a: AnyFile::Std(
-                StdFile::open(sidecar_path(path, ".wal-a")).map_err(map_storage_err)?,
-            ),
-            wal_b: AnyFile::Std(
-                StdFile::open(sidecar_path(path, ".wal-b")).map_err(map_storage_err)?,
-            ),
-            manifest: AnyFile::Std(
-                StdFile::open(sidecar_path(path, ".manifest")).map_err(map_storage_err)?,
-            ),
-        })
-    }
-
-    #[cfg(target_arch = "wasm32")]
-    async fn open_opfs(namespace: &str) -> Result<Self, StorageError> {
-        Ok(Self {
-            snapshot_a: AnyFile::Opfs(OpfsFile::open(namespace).await.map_err(map_storage_err)?),
-            snapshot_b: AnyFile::Opfs(
-                OpfsFile::open(&opfs_sidecar_name(namespace, "snapshot-b"))
-                    .await
-                    .map_err(map_storage_err)?,
-            ),
-            wal_a: AnyFile::Opfs(
-                OpfsFile::open(&opfs_sidecar_name(namespace, "wal-a"))
-                    .await
-                    .map_err(map_storage_err)?,
-            ),
-            wal_b: AnyFile::Opfs(
-                OpfsFile::open(&opfs_sidecar_name(namespace, "wal-b"))
-                    .await
-                    .map_err(map_storage_err)?,
-            ),
-            manifest: AnyFile::Opfs(
-                OpfsFile::open(&opfs_sidecar_name(namespace, "manifest"))
-                    .await
-                    .map_err(map_storage_err)?,
-            ),
-        })
-    }
-
-    fn snapshot(&self, slot: ReplicaSlot) -> AnyFile {
-        match slot {
-            ReplicaSlot::A => self.snapshot_a.clone(),
-            ReplicaSlot::B => self.snapshot_b.clone(),
-        }
-    }
-
-    fn wal(&self, slot: ReplicaSlot) -> AnyFile {
-        match slot {
-            ReplicaSlot::A => self.wal_a.clone(),
-            ReplicaSlot::B => self.wal_b.clone(),
-        }
-    }
-}
-
-#[derive(Debug)]
-struct StorageState {
-    files: DurableFiles,
-    options: BTreeOptions,
-    tree: OpfsBTree<AnyFile>,
-    manifest_slot: ManifestSlot,
-    manifest: DurableManifest,
-    pending_wal: Vec<WalEntry>,
-    wal_flushed_seq: u64,
-    next_wal_seq: u64,
-}
-
-#[cfg(test)]
-mod test_failpoints {
-    use std::cell::RefCell;
-
-    use super::StorageError;
-
-    #[derive(Default)]
-    struct State {
-        armed_step: Option<usize>,
-        next_step: usize,
-        hit_sites: Vec<&'static str>,
-    }
-
-    std::thread_local! {
-        static STATE: RefCell<State> = RefCell::new(State::default());
-    }
-
-    pub(crate) fn arm(step: usize) {
-        STATE.with(|cell| {
-            let mut state = cell.borrow_mut();
-            state.armed_step = Some(step);
-            state.next_step = 0;
-            state.hit_sites.clear();
-        });
-    }
-
-    pub(crate) fn clear() {
-        STATE.with(|cell| {
-            let mut state = cell.borrow_mut();
-            state.armed_step = None;
-            state.next_step = 0;
-            state.hit_sites.clear();
-        });
-    }
-
-    pub(crate) fn step_count() -> usize {
-        STATE.with(|cell| cell.borrow().next_step)
-    }
-
-    pub(crate) fn hit_sites() -> Vec<&'static str> {
-        STATE.with(|cell| cell.borrow().hit_sites.clone())
-    }
-
-    pub(crate) fn hit(site: &'static str) -> Result<(), StorageError> {
-        STATE.with(|cell| {
-            let mut state = cell.borrow_mut();
-            state.next_step = state.next_step.saturating_add(1);
-            let step = state.next_step;
-            state.hit_sites.push(site);
-            if state.armed_step == Some(step) {
-                return Err(StorageError::IoError(format!(
-                    "opfs-btree durable failpoint at step {} ({})",
-                    step, site
-                )));
-            }
-            Ok(())
-        })
-    }
-}
-
-#[cfg(test)]
-macro_rules! durable_failpoint {
-    ($site:literal) => {
-        test_failpoints::hit($site)?;
-    };
-}
-
-#[cfg(not(test))]
-macro_rules! durable_failpoint {
-    ($site:literal) => {};
-}
-
 pub struct OpfsBTreeStorage {
-    state: RefCell<StorageState>,
+    tree: RefCell<DurableOpfsBTree<AnyFile>>,
 }
 
 impl OpfsBTreeStorage {
     #[cfg(not(target_arch = "wasm32"))]
     pub fn open(path: impl AsRef<Path>, cache_size_bytes: usize) -> Result<Self, StorageError> {
-        let files = DurableFiles::open(path)?;
+        let files = DurableBTreeFiles::<StdFile>::open(path)
+            .map(|files| files.map(AnyFile::Std))
+            .map_err(map_storage_err)?;
         Self::open_with_files(files, cache_size_bytes)
     }
 
     pub fn memory(cache_size_bytes: usize) -> Result<Self, StorageError> {
-        Self::open_with_files(DurableFiles::memory(), cache_size_bytes)
+        let files = DurableBTreeFiles::<MemoryFile>::memory().map(AnyFile::Memory);
+        Self::open_with_files(files, cache_size_bytes)
     }
 
     #[cfg(target_arch = "wasm32")]
@@ -400,53 +136,28 @@ impl OpfsBTreeStorage {
 
     #[cfg(target_arch = "wasm32")]
     pub async fn open_opfs(namespace: &str, cache_size_bytes: usize) -> Result<Self, StorageError> {
-        let files = DurableFiles::open_opfs(namespace).await?;
+        let files = DurableBTreeFiles::<OpfsFile>::open_opfs(namespace)
+            .await
+            .map_err(map_storage_err)?
+            .map(AnyFile::Opfs);
         Self::open_with_files(files, cache_size_bytes)
     }
 
     #[cfg(target_arch = "wasm32")]
     pub async fn destroy_opfs(namespace: &str) -> Result<(), StorageError> {
-        OpfsFile::destroy(namespace)
+        DurableBTreeFiles::<OpfsFile>::destroy_opfs(namespace)
             .await
-            .map_err(map_storage_err)?;
-        let _ = OpfsFile::destroy(&opfs_sidecar_name(namespace, "snapshot-b")).await;
-        let _ = OpfsFile::destroy(&opfs_sidecar_name(namespace, "wal-a")).await;
-        let _ = OpfsFile::destroy(&opfs_sidecar_name(namespace, "wal-b")).await;
-        let _ = OpfsFile::destroy(&opfs_sidecar_name(namespace, "manifest")).await;
-        Ok(())
+            .map_err(map_storage_err)
     }
 
-    fn open_with_files(files: DurableFiles, cache_size_bytes: usize) -> Result<Self, StorageError> {
-        let options = Self::options(cache_size_bytes);
-        let (manifest_slot, manifest) = read_manifest(&files.manifest)?
-            .unwrap_or((ManifestSlot::A, DurableManifest::bootstrap()));
-        durable_failpoint!("open:after-read-manifest");
-
-        let mut tree = OpfsBTree::open(files.snapshot(manifest.active_snapshot), options)
+    fn open_with_files(
+        files: DurableBTreeFiles<AnyFile>,
+        cache_size_bytes: usize,
+    ) -> Result<Self, StorageError> {
+        let tree = DurableOpfsBTree::open(files, Self::options(cache_size_bytes))
             .map_err(map_storage_err)?;
-        durable_failpoint!("open:after-open-snapshot");
-
-        let wal_entries = read_wal_entries(&files.wal(manifest.active_wal))?;
-        let mut wal_flushed_seq = manifest.applied_wal_seq;
-        for entry in wal_entries {
-            wal_flushed_seq = wal_flushed_seq.max(entry.seq());
-            if entry.seq() > manifest.applied_wal_seq {
-                apply_wal_entry(&mut tree, &entry)?;
-            }
-        }
-        durable_failpoint!("open:after-replay-wal");
-
         Ok(Self {
-            state: RefCell::new(StorageState {
-                files,
-                options,
-                tree,
-                manifest_slot,
-                manifest,
-                pending_wal: Vec::new(),
-                wal_flushed_seq,
-                next_wal_seq: wal_flushed_seq.saturating_add(1),
-            }),
+            tree: RefCell::new(tree),
         })
     }
 
@@ -459,27 +170,27 @@ impl OpfsBTreeStorage {
         }
     }
 
-    fn with_state_mut<R>(
+    fn with_tree_mut<R>(
         &self,
-        f: impl FnOnce(&mut StorageState) -> Result<R, StorageError>,
+        f: impl FnOnce(&mut DurableOpfsBTree<AnyFile>) -> Result<R, StorageError>,
     ) -> Result<R, StorageError> {
-        let mut state = self
-            .state
+        let mut tree = self
+            .tree
             .try_borrow_mut()
             .map_err(|_| StorageError::IoError("opfs-btree already borrowed".to_string()))?;
-        f(&mut state)
+        f(&mut tree)
     }
 
     fn tree_insert(&self, key: &str, value: &[u8]) -> Result<(), StorageError> {
-        self.with_state_mut(|state| state.put(key.as_bytes(), value))
+        self.with_tree_mut(|tree| tree.put(key.as_bytes(), value).map_err(map_storage_err))
     }
 
     fn tree_read(&self, key: &str) -> Result<Option<Vec<u8>>, StorageError> {
-        self.with_state_mut(|state| state.tree.get(key.as_bytes()).map_err(map_storage_err))
+        self.with_tree_mut(|tree| tree.get(key.as_bytes()).map_err(map_storage_err))
     }
 
     fn tree_delete(&self, key: &str) -> Result<(), StorageError> {
-        self.with_state_mut(|state| state.delete(key.as_bytes()))
+        self.with_tree_mut(|tree| tree.delete(key.as_bytes()).map_err(map_storage_err))
     }
 
     fn tree_scan_prefix(&self, prefix: &str) -> Result<Vec<(String, Vec<u8>)>, StorageError> {
@@ -514,9 +225,8 @@ impl OpfsBTreeStorage {
             return Ok(Vec::new());
         }
 
-        self.with_state_mut(|state| {
-            let entries = state
-                .tree
+        self.with_tree_mut(|tree| {
+            let entries = tree
                 .range(start, end, usize::MAX)
                 .map_err(map_storage_err)?;
 
@@ -530,429 +240,6 @@ impl OpfsBTreeStorage {
                 .collect()
         })
     }
-}
-
-impl StorageState {
-    fn put(&mut self, key: &[u8], value: &[u8]) -> Result<(), StorageError> {
-        self.tree.put(key, value).map_err(map_storage_err)?;
-        let seq = self.alloc_seq();
-        self.pending_wal.push(WalEntry::Put {
-            seq,
-            key: key.to_vec(),
-            value: value.to_vec(),
-        });
-        Ok(())
-    }
-
-    fn delete(&mut self, key: &[u8]) -> Result<(), StorageError> {
-        self.tree.delete(key).map_err(map_storage_err)?;
-        let seq = self.alloc_seq();
-        self.pending_wal.push(WalEntry::Delete {
-            seq,
-            key: key.to_vec(),
-        });
-        Ok(())
-    }
-
-    fn alloc_seq(&mut self) -> u64 {
-        let seq = self.next_wal_seq;
-        self.next_wal_seq = self.next_wal_seq.saturating_add(1);
-        seq
-    }
-
-    fn flush_wal_inner(&mut self) -> Result<(), StorageError> {
-        if self.pending_wal.is_empty() {
-            return Ok(());
-        }
-
-        let active_wal = self.files.wal(self.manifest.active_wal);
-        let batch = encode_wal_batch(&self.pending_wal)?;
-        let offset = active_wal.len().map_err(map_storage_err)?;
-        durable_failpoint!("flush_wal:before-write");
-        active_wal
-            .write_all_at(offset, &batch)
-            .map_err(map_storage_err)?;
-        durable_failpoint!("flush_wal:after-write");
-        active_wal.flush().map_err(map_storage_err)?;
-        durable_failpoint!("flush_wal:after-flush");
-
-        self.wal_flushed_seq = self
-            .pending_wal
-            .last()
-            .map(WalEntry::seq)
-            .unwrap_or(self.wal_flushed_seq);
-        self.pending_wal.clear();
-        Ok(())
-    }
-
-    fn checkpoint_inner(&mut self) -> Result<(), StorageError> {
-        self.flush_wal_inner()?;
-        durable_failpoint!("checkpoint:after-flush-wal");
-
-        if self.wal_flushed_seq == self.manifest.applied_wal_seq {
-            return Ok(());
-        }
-
-        let target_snapshot = self.manifest.active_snapshot.inactive();
-        let target_wal = self.manifest.active_wal.inactive();
-
-        let snapshot_file = self.files.snapshot(target_snapshot);
-        snapshot_file.truncate(0).map_err(map_storage_err)?;
-        durable_failpoint!("checkpoint:after-truncate-snapshot");
-
-        let mut snapshot_tree =
-            OpfsBTree::open(snapshot_file, self.options).map_err(map_storage_err)?;
-        durable_failpoint!("checkpoint:after-open-target-snapshot");
-
-        for (key, value) in self.full_scan()? {
-            snapshot_tree.put(&key, &value).map_err(map_storage_err)?;
-        }
-        durable_failpoint!("checkpoint:after-build-snapshot");
-        snapshot_tree.checkpoint().map_err(map_storage_err)?;
-        durable_failpoint!("checkpoint:after-snapshot-checkpoint");
-
-        let next_wal_file = self.files.wal(target_wal);
-        next_wal_file.truncate(0).map_err(map_storage_err)?;
-        next_wal_file.flush().map_err(map_storage_err)?;
-        durable_failpoint!("checkpoint:after-prepare-inactive-wal");
-
-        let next_manifest = DurableManifest {
-            generation: self.manifest.generation.saturating_add(1),
-            active_snapshot: target_snapshot,
-            active_wal: target_wal,
-            applied_wal_seq: self.wal_flushed_seq,
-        };
-        let next_manifest_slot = self.manifest_slot.inactive();
-        write_manifest(&self.files.manifest, next_manifest_slot, next_manifest)?;
-        durable_failpoint!("checkpoint:after-manifest-write");
-        self.files.manifest.flush().map_err(map_storage_err)?;
-        durable_failpoint!("checkpoint:after-manifest-flush");
-
-        self.tree = snapshot_tree;
-        self.manifest = next_manifest;
-        self.manifest_slot = next_manifest_slot;
-        self.pending_wal.clear();
-        durable_failpoint!("checkpoint:after-tree-swap");
-        Ok(())
-    }
-
-    fn full_scan(&mut self) -> Result<Vec<EncodedKv>, StorageError> {
-        self.tree
-            .range(b"", &FULL_SCAN_END_KEY, usize::MAX)
-            .map_err(map_storage_err)
-    }
-}
-
-#[cfg(not(target_arch = "wasm32"))]
-fn sidecar_path(path: &Path, suffix: &str) -> PathBuf {
-    let file_name = path
-        .file_name()
-        .map(|value| value.to_string_lossy().into_owned())
-        .unwrap_or_else(|| "jazz.opfsbtree".to_string());
-    let mut out = path.to_path_buf();
-    out.set_file_name(format!("{}{}", file_name, suffix));
-    out
-}
-
-#[cfg(target_arch = "wasm32")]
-fn opfs_sidecar_name(namespace: &str, suffix: &str) -> String {
-    format!("{namespace}__{suffix}")
-}
-
-fn apply_wal_entry(tree: &mut OpfsBTree<AnyFile>, entry: &WalEntry) -> Result<(), StorageError> {
-    match entry {
-        WalEntry::Put { key, value, .. } => tree.put(key, value).map_err(map_storage_err),
-        WalEntry::Delete { key, .. } => tree.delete(key).map_err(map_storage_err),
-    }
-}
-
-fn manifest_checksum(bytes: &[u8]) -> [u8; MANIFEST_CHECKSUM_BYTES] {
-    let hash = blake3::hash(bytes);
-    let mut out = [0u8; MANIFEST_CHECKSUM_BYTES];
-    out.copy_from_slice(&hash.as_bytes()[..MANIFEST_CHECKSUM_BYTES]);
-    out
-}
-
-fn encode_manifest(manifest: DurableManifest) -> [u8; MANIFEST_SLOT_BYTES] {
-    let mut slot = [0u8; MANIFEST_SLOT_BYTES];
-    let mut payload = Vec::with_capacity(32);
-    payload.extend_from_slice(&MANIFEST_MAGIC);
-    payload.extend_from_slice(&MANIFEST_VERSION.to_le_bytes());
-    payload.extend_from_slice(&manifest.generation.to_le_bytes());
-    payload.push(manifest.active_snapshot.encode());
-    payload.push(manifest.active_wal.encode());
-    payload.extend_from_slice(&manifest.applied_wal_seq.to_le_bytes());
-    let checksum = manifest_checksum(&payload);
-    let payload_len = payload.len();
-    slot[..payload_len].copy_from_slice(&payload);
-    slot[payload_len..payload_len + MANIFEST_CHECKSUM_BYTES].copy_from_slice(&checksum);
-    slot
-}
-
-fn decode_manifest(slot: &[u8]) -> Result<DurableManifest, StorageError> {
-    if slot.len() < MANIFEST_SLOT_BYTES {
-        return Err(StorageError::IoError(
-            "opfs-btree durable manifest slot too small".to_string(),
-        ));
-    }
-    if slot[..MANIFEST_MAGIC.len()] != MANIFEST_MAGIC {
-        return Err(StorageError::IoError(
-            "opfs-btree durable manifest magic mismatch".to_string(),
-        ));
-    }
-
-    let checksum_offset = 8 + 4 + 8 + 1 + 1 + 8;
-    let expected_checksum = manifest_checksum(&slot[..checksum_offset]);
-    let actual_checksum = &slot[checksum_offset..checksum_offset + MANIFEST_CHECKSUM_BYTES];
-    if actual_checksum != expected_checksum {
-        return Err(StorageError::IoError(
-            "opfs-btree durable manifest checksum mismatch".to_string(),
-        ));
-    }
-
-    let version = u32::from_le_bytes(
-        slot[8..12]
-            .try_into()
-            .expect("manifest version slice must fit"),
-    );
-    if version != MANIFEST_VERSION {
-        return Err(StorageError::IoError(format!(
-            "opfs-btree durable manifest version {} is unsupported",
-            version
-        )));
-    }
-
-    let generation = u64::from_le_bytes(
-        slot[12..20]
-            .try_into()
-            .expect("manifest generation slice must fit"),
-    );
-    let active_snapshot = ReplicaSlot::decode(slot[20])?;
-    let active_wal = ReplicaSlot::decode(slot[21])?;
-    let applied_wal_seq = u64::from_le_bytes(
-        slot[22..30]
-            .try_into()
-            .expect("manifest wal seq slice must fit"),
-    );
-
-    Ok(DurableManifest {
-        generation,
-        active_snapshot,
-        active_wal,
-        applied_wal_seq,
-    })
-}
-
-fn read_manifest(file: &AnyFile) -> Result<Option<(ManifestSlot, DurableManifest)>, StorageError> {
-    if file.len().map_err(map_storage_err)? < MANIFEST_SLOT_BYTES as u64 {
-        return Ok(None);
-    }
-
-    let mut best: Option<(ManifestSlot, DurableManifest)> = None;
-    for slot_id in [ManifestSlot::A, ManifestSlot::B] {
-        let needed = slot_id.byte_offset() + MANIFEST_SLOT_BYTES as u64;
-        if file.len().map_err(map_storage_err)? < needed {
-            continue;
-        }
-
-        let mut raw = [0u8; MANIFEST_SLOT_BYTES];
-        file.read_exact_at(slot_id.byte_offset(), &mut raw)
-            .map_err(map_storage_err)?;
-        let Ok(manifest) = decode_manifest(&raw) else {
-            continue;
-        };
-        match best {
-            Some((_, current)) if current.generation >= manifest.generation => {}
-            _ => best = Some((slot_id, manifest)),
-        }
-    }
-
-    Ok(best)
-}
-
-fn write_manifest(
-    file: &AnyFile,
-    slot: ManifestSlot,
-    manifest: DurableManifest,
-) -> Result<(), StorageError> {
-    let raw = encode_manifest(manifest);
-    file.write_all_at(slot.byte_offset(), &raw)
-        .map_err(map_storage_err)
-}
-
-fn wal_checksum(bytes: &[u8]) -> [u8; MANIFEST_CHECKSUM_BYTES] {
-    manifest_checksum(bytes)
-}
-
-fn encode_wal_batch(entries: &[WalEntry]) -> Result<Vec<u8>, StorageError> {
-    let mut body = Vec::new();
-    let start_seq = entries
-        .first()
-        .map(WalEntry::seq)
-        .ok_or_else(|| StorageError::IoError("empty WAL batch".to_string()))?;
-    body.extend_from_slice(&start_seq.to_le_bytes());
-    body.extend_from_slice(
-        &(u32::try_from(entries.len())
-            .map_err(|_| StorageError::IoError("WAL batch too large".to_string()))?)
-        .to_le_bytes(),
-    );
-
-    for entry in entries {
-        match entry {
-            WalEntry::Put { key, value, .. } => {
-                body.push(1);
-                body.extend_from_slice(
-                    &(u32::try_from(key.len())
-                        .map_err(|_| StorageError::IoError("WAL key too large".to_string()))?)
-                    .to_le_bytes(),
-                );
-                body.extend_from_slice(
-                    &(u32::try_from(value.len())
-                        .map_err(|_| StorageError::IoError("WAL value too large".to_string()))?)
-                    .to_le_bytes(),
-                );
-                body.extend_from_slice(key);
-                body.extend_from_slice(value);
-            }
-            WalEntry::Delete { key, .. } => {
-                body.push(2);
-                body.extend_from_slice(
-                    &(u32::try_from(key.len())
-                        .map_err(|_| StorageError::IoError("WAL key too large".to_string()))?)
-                    .to_le_bytes(),
-                );
-                body.extend_from_slice(&0u32.to_le_bytes());
-                body.extend_from_slice(key);
-            }
-        }
-    }
-
-    let checksum = wal_checksum(&body);
-    let mut out = Vec::with_capacity(WAL_BATCH_HEADER_BYTES + body.len());
-    out.extend_from_slice(&WAL_BATCH_MAGIC);
-    out.extend_from_slice(
-        &(u32::try_from(body.len())
-            .map_err(|_| StorageError::IoError("WAL batch body too large".to_string()))?)
-        .to_le_bytes(),
-    );
-    out.extend_from_slice(&checksum);
-    out.extend_from_slice(&body);
-    Ok(out)
-}
-
-fn decode_wal_batch(bytes: &[u8]) -> Result<Vec<WalEntry>, StorageError> {
-    if bytes.len() < 12 {
-        return Err(StorageError::IoError(
-            "WAL batch body too small".to_string(),
-        ));
-    }
-
-    let start_seq = u64::from_le_bytes(
-        bytes[..8]
-            .try_into()
-            .expect("WAL batch start seq slice must fit"),
-    );
-    let entry_count = u32::from_le_bytes(
-        bytes[8..12]
-            .try_into()
-            .expect("WAL batch count slice must fit"),
-    ) as usize;
-
-    let mut entries = Vec::with_capacity(entry_count);
-    let mut offset = 12usize;
-    for index in 0..entry_count {
-        if offset + 9 > bytes.len() {
-            return Err(StorageError::IoError(
-                "WAL batch entry header truncated".to_string(),
-            ));
-        }
-        let op = bytes[offset];
-        offset += 1;
-        let key_len = u32::from_le_bytes(
-            bytes[offset..offset + 4]
-                .try_into()
-                .expect("WAL key length slice must fit"),
-        ) as usize;
-        offset += 4;
-        let value_len = u32::from_le_bytes(
-            bytes[offset..offset + 4]
-                .try_into()
-                .expect("WAL value length slice must fit"),
-        ) as usize;
-        offset += 4;
-        if offset + key_len > bytes.len() {
-            return Err(StorageError::IoError("WAL key truncated".to_string()));
-        }
-        let key = bytes[offset..offset + key_len].to_vec();
-        offset += key_len;
-        let seq = start_seq.saturating_add(index as u64);
-        match op {
-            1 => {
-                if offset + value_len > bytes.len() {
-                    return Err(StorageError::IoError("WAL value truncated".to_string()));
-                }
-                let value = bytes[offset..offset + value_len].to_vec();
-                offset += value_len;
-                entries.push(WalEntry::Put { seq, key, value });
-            }
-            2 => entries.push(WalEntry::Delete { seq, key }),
-            other => {
-                return Err(StorageError::IoError(format!(
-                    "WAL operation {} is invalid",
-                    other
-                )));
-            }
-        }
-    }
-
-    if offset != bytes.len() {
-        return Err(StorageError::IoError(
-            "WAL batch trailing bytes are invalid".to_string(),
-        ));
-    }
-    Ok(entries)
-}
-
-fn read_wal_entries(file: &AnyFile) -> Result<Vec<WalEntry>, StorageError> {
-    let len = usize::try_from(file.len().map_err(map_storage_err)?)
-        .map_err(|_| StorageError::IoError("WAL length does not fit in usize".to_string()))?;
-    if len == 0 {
-        return Ok(Vec::new());
-    }
-
-    let mut raw = vec![0u8; len];
-    file.read_exact_at(0, &mut raw).map_err(map_storage_err)?;
-
-    let mut entries = Vec::new();
-    let mut offset = 0usize;
-    while offset + WAL_BATCH_HEADER_BYTES <= raw.len() {
-        if raw[offset..offset + 8] != WAL_BATCH_MAGIC {
-            break;
-        }
-        let body_len = u32::from_le_bytes(
-            raw[offset + 8..offset + 12]
-                .try_into()
-                .expect("WAL batch length slice must fit"),
-        ) as usize;
-        let batch_end = offset + WAL_BATCH_HEADER_BYTES + body_len;
-        if batch_end > raw.len() {
-            break;
-        }
-
-        let expected_checksum = wal_checksum(&raw[offset + WAL_BATCH_HEADER_BYTES..batch_end]);
-        let checksum = &raw[offset + 12..offset + 12 + MANIFEST_CHECKSUM_BYTES];
-        if checksum != expected_checksum {
-            break;
-        }
-
-        let Ok(batch) = decode_wal_batch(&raw[offset + WAL_BATCH_HEADER_BYTES..batch_end]) else {
-            break;
-        };
-        entries.extend(batch);
-        offset = batch_end;
-    }
-
-    Ok(entries)
 }
 
 impl Storage for OpfsBTreeStorage {
@@ -1136,14 +423,14 @@ impl Storage for OpfsBTreeStorage {
 
     fn flush(&self) {
         let _span = tracing::debug_span!("OpfsBTreeStorage::flush").entered();
-        if let Err(error) = self.with_state_mut(|state| state.checkpoint_inner()) {
+        if let Err(error) = self.with_tree_mut(|tree| tree.checkpoint().map_err(map_storage_err)) {
             tracing::warn!(?error, "OpfsBTreeStorage flush failed");
         }
     }
 
     fn flush_wal(&self) {
         let _span = tracing::debug_span!("OpfsBTreeStorage::flush_wal").entered();
-        if let Err(error) = self.with_state_mut(|state| state.flush_wal_inner()) {
+        if let Err(error) = self.with_tree_mut(|tree| tree.flush_wal().map_err(map_storage_err)) {
             tracing::warn!(?error, "OpfsBTreeStorage WAL flush failed");
         }
     }
@@ -1157,7 +444,6 @@ fn map_storage_err(error: BTreeError) -> StorageError {
 mod tests {
     use super::*;
     use smallvec::smallvec;
-    use std::collections::BTreeMap;
 
     fn make_commit(content: &[u8]) -> Commit {
         Commit {
@@ -1173,56 +459,6 @@ mod tests {
 
     fn test_storage() -> OpfsBTreeStorage {
         OpfsBTreeStorage::memory(4 * 1024 * 1024).unwrap()
-    }
-
-    fn durable_fixture() -> (DurableFiles, usize) {
-        (DurableFiles::memory(), 4 * 1024 * 1024)
-    }
-
-    fn open_fixture(files: DurableFiles, cache_size_bytes: usize) -> OpfsBTreeStorage {
-        OpfsBTreeStorage::open_with_files(files, cache_size_bytes).expect("open fixture storage")
-    }
-
-    fn apply_baseline(storage: &OpfsBTreeStorage) {
-        storage
-            .tree_insert("kv:stable", b"old")
-            .expect("insert stable");
-        storage
-            .tree_insert("kv:delete-me", b"remove-later")
-            .expect("insert delete-me");
-        storage
-            .with_state_mut(|state| state.checkpoint_inner())
-            .expect("checkpoint baseline");
-    }
-
-    fn apply_update(storage: &OpfsBTreeStorage) {
-        storage
-            .tree_insert("kv:stable", b"new")
-            .expect("update stable");
-        storage.tree_insert("kv:new", b"fresh").expect("insert new");
-        storage.tree_delete("kv:delete-me").expect("delete key");
-    }
-
-    fn collect_tree(storage: &OpfsBTreeStorage) -> BTreeMap<String, Vec<u8>> {
-        storage
-            .tree_scan_range_bytes(b"", &FULL_SCAN_END_KEY)
-            .expect("scan full tree")
-            .into_iter()
-            .collect()
-    }
-
-    fn baseline_tree() -> BTreeMap<String, Vec<u8>> {
-        BTreeMap::from([
-            ("kv:delete-me".to_string(), b"remove-later".to_vec()),
-            ("kv:stable".to_string(), b"old".to_vec()),
-        ])
-    }
-
-    fn updated_tree() -> BTreeMap<String, Vec<u8>> {
-        BTreeMap::from([
-            ("kv:new".to_string(), b"fresh".to_vec()),
-            ("kv:stable".to_string(), b"new".to_vec()),
-        ])
     }
 
     #[test]
@@ -1466,77 +702,6 @@ mod tests {
             assert_eq!(
                 storage.tree_read("kv:stable").unwrap(),
                 Some(b"wal-value".to_vec())
-            );
-        }
-    }
-
-    #[test]
-    fn durable_failpoint_sweep_requires_clean_recovery_for_every_step() {
-        test_failpoints::clear();
-        let cache_size_bytes = 4 * 1024 * 1024;
-        let (probe_files, _) = durable_fixture();
-
-        let max_step = {
-            let storage = open_fixture(probe_files.clone(), cache_size_bytes);
-            apply_baseline(&storage);
-            apply_update(&storage);
-            test_failpoints::clear();
-            storage
-                .with_state_mut(|state| state.checkpoint_inner())
-                .expect("baseline durable checkpoint");
-            test_failpoints::step_count()
-        };
-        assert!(max_step > 0, "expected at least one durable failpoint");
-        test_failpoints::clear();
-
-        let baseline = baseline_tree();
-        let updated = updated_tree();
-        let mut failures = Vec::new();
-
-        for step in 1..=max_step {
-            let (files, _) = durable_fixture();
-            let storage = open_fixture(files.clone(), cache_size_bytes);
-            apply_baseline(&storage);
-            drop(storage);
-
-            let storage = open_fixture(files.clone(), cache_size_bytes);
-            apply_update(&storage);
-            test_failpoints::arm(step);
-            let result = storage.with_state_mut(|state| state.checkpoint_inner());
-            let hit_sites = test_failpoints::hit_sites();
-            let site = hit_sites
-                .get(step.saturating_sub(1))
-                .copied()
-                .or_else(|| hit_sites.last().copied())
-                .unwrap_or("<no-site-recorded>");
-            match result {
-                Ok(()) => failures.push(format!(
-                    "step {step}: {site} -> armed failpoint did not abort durable checkpoint"
-                )),
-                Err(StorageError::IoError(message)) if message.contains("durable failpoint") => {}
-                Err(error) => failures.push(format!(
-                    "step {step}: {site} -> unexpected error during checkpoint: {:?}",
-                    error
-                )),
-            }
-            drop(storage);
-            test_failpoints::clear();
-
-            let reopened = open_fixture(files.clone(), cache_size_bytes);
-            let recovered = collect_tree(&reopened);
-            if recovered != baseline && recovered != updated {
-                failures.push(format!(
-                    "step {step}: {site} -> mixed durable recovery state: {:?}",
-                    recovered
-                ));
-            }
-        }
-
-        if !failures.is_empty() {
-            panic!(
-                "durable failpoint sweep found {} failing step(s)\n{}",
-                failures.len(),
-                failures.join("\n")
             );
         }
     }
