+++ crates/opfs-btree/src/lib.rs
@@ -1,4 +1,5 @@
 mod db;
+mod durable;
 mod error;
 mod file;
 mod page;
@@ -7,6 +8,7 @@ mod superblock;
 pub mod wasm_bench;
 
 pub use db::{BTreeOptions, CheckpointState, OpfsBTree};
+pub use durable::{DurableBTreeFiles, DurableOpfsBTree};
 pub use error::BTreeError;
 #[cfg(target_arch = "wasm32")]
 pub use file::OpfsFile;
